+++ src/main/java/net/sf/jabref/gui/GenFieldsCustomizer.java
@@ -139,7 +139,6 @@ public class GenFieldsCustomizer extends JDialog {
         Globals.prefs.purgeSeries(JabRefPreferences.CUSTOM_TAB_FIELDS, i);
         Globals.prefs.updateEntryEditorTabList();
 
-        parentFrame.removeCachedEntryEditors();
         dispose();
     }