+++ src/main/java/net/sf/jabref/logic/formatter/bibtexfields/ClearFormatter.java
@@ -2,8 +2,8 @@ package net.sf.jabref.logic.formatter.bibtexfields;
 
 import java.util.Objects;
 
-import net.sf.jabref.logic.formatter.Formatter;
 import net.sf.jabref.logic.l10n.Localization;
+import net.sf.jabref.model.cleanup.Formatter;
 
 public class ClearFormatter implements Formatter {