+++ src/main/java/net/sf/jabref/logic/importer/SearchBasedParserFetcher.java
@@ -9,7 +9,7 @@ import java.net.URL;
 import java.util.Collections;
 import java.util.List;
 
-import net.sf.jabref.logic.formatter.Formatter;
+import net.sf.jabref.model.cleanup.Formatter;
 import net.sf.jabref.model.entry.BibEntry;
 
 import org.jsoup.helper.StringUtil;
@@ -68,7 +68,7 @@ public interface SearchBasedParserFetcher extends SearchBasedFetcher {
         } catch (IOException e) {
             // TODO: Catch HTTP Response 401 errors and report that user has no rights to access resource
             throw new FetcherException("An I/O exception occurred", e);
-        } catch (ParserException e) {
+        } catch (ParseException e) {
             throw new FetcherException("An internal parser error occurred", e);
         }
     }