+++ src/main/java/net/sf/jabref/logic/importer/fileformat/SilverPlatterImporter.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.regex.Pattern;
 
+import net.sf.jabref.logic.importer.Importer;
 import net.sf.jabref.logic.importer.ParserResult;
 import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.model.entry.AuthorList;
@@ -18,12 +19,12 @@ import net.sf.jabref.model.entry.FieldName;
  * Imports a SilverPlatter exported file. This is a poor format to parse,
  * so it currently doesn't handle everything correctly.
  */
-public class SilverPlatterImporter extends ImportFormat {
+public class SilverPlatterImporter extends Importer {
 
     private static final Pattern START_PATTERN = Pattern.compile("Record.*INSPEC.*");
 
     @Override
-    public String getFormatName() {
+    public String getName() {
         return "SilverPlatter";
     }